+++ CHANGELOG.md
@@ -17,6 +17,8 @@ Any non-code changes should be prefixed with `(docs)`.
 See `PUBLISH.md` for instructions on how to publish a new version.
 -->
 
+- (minor) Add Image Compare embeds
+- (minor) Add Slideshow embeds
 - (minor) Add Instagram embeds
 - (minor) Add Vimeo embeds